+++ Tests/Functions/Private/ConvertTo-JiraIssueLinkType.Unit.Tests.ps1
@@ -3,10 +3,7 @@
 BeforeDiscovery {
     . "$PSScriptRoot/../../Helpers/TestTools.ps1"
 
-    Initialize-TestEnvironment
-    $script:moduleToTest = Resolve-ModuleSource
-
-    Import-Module $moduleToTest -Force -ErrorAction Stop
+    $script:moduleToTest = Initialize-TestEnvironment
 }
 
 InModuleScope JiraPS {